Migrate query parameters to own sub directory
This also combines handling of schema, and logic for each query parameter
