--- scripts/parameters/auth.js
@@ -1,10 +1,0 @@
-module.exports = {
-  name: "auth",
-  in: "header",
-  schema: {
-    type: "string"
-  },
-  required: true,
-  allowEmptyValue: false,
-  description: "Authorization Headers."
-};
