--- scripts/parameters/login.js
@@ -1,11 +1,0 @@
-module.exports = {
-  name: "login",
-  in: "path",
-  schema: {
-    type: "string"
-  },
-  required: true,
-  allowEmptyValue: false,
-  example: "confused-Techie",
-  description: "The User from the URL path."
-};
