--- scripts/parameters/query.js
@@ -1,10 +1,0 @@
-module.exports = {
-  name: "q",
-  in: "query",
-  schema: {
-    type: "string"
-  },
-  example: "generic-lsp",
-  required: true,
-  description: "Search query."
-};
