--- scripts/parameters/service.js
@@ -1,10 +1,0 @@
-module.exports = {
-  name: "service",
-  in: "query",
-  schema: {
-    type: "string"
-  },
-  example: "autocomplete.watchEditor",
-  allowEmptyValue: true,
-  description: "The service of which to filter packages by."
-};
